Get rid of s-provide-complete completely
All provides are now desugared to the new s-provide-block

hiding and module references are not currently supported in provides, nor is
the new design for re-providing from an imported module

In addition, there's a funky thing going on in resolve-scope to make this work
in its current state: If *anything* is provided from a module as a value or a
type, then *all* the datatypes are included.

This is actually not particularly different than what happens now (where all
datatypes are always exposed, BUT never added to a namespace on include). It's
just a weirdness of the implementation.

The long-term solution here is to figure out what the different modalities of
providing a datatype are (e.g. provide it opaquely, hide its constructors, and
so on) and how those interact with the value & type namespaces