Fix exports to actually provide all names, and re-enable error
There were several real typos or references to undefined constructors here!

In addition, a lot of js/trove didn't list all exports.

Finally, add-module-name was using the URI of the module doing the including,
not the URI of the module that was referred to, when binding module-ids, so fix
that.