Adding useful description to last commit, and forgotten files
The new idea here is that bind-origin also tracks the original name of a
value/type/module binding.

This turns out to be necessary, because without this there was no place to
talk about the following scenario. Consider:

    provide: sd as str-dict end
    import string-dict as SD
    include from SD: string-dict as sd end

When providing sd, it needs to be that str-dict is an *alias* for string-dict
on the builtin://string-dict module. Previously, there was no way for the
compiler to determine which name it was supposed to be an alias for – based on
the bind-origin it would know that it was supposed to be found on
builtin://string-dict, but nothing tracked the original name in the
module-local environment.

This idea is pretty powerful. Among other things, it likely means that v-alias
can lose the "original-name" field. I also think the original-name field can
probably be a String, because there's little value to it holding an atom when
the atom is only used locally (when provided, the pair of module URI plus plain
string name is enough).

I think that this could simplify some other positions, and there are likely
other helpers on CompileEnvironment that can be written as a result. There are
extra checks in a few lookup locations that use original-name now, which should
probably be encapsulated.