Fix an issue with reassigning parameters in TRO calls
[Reported via email]

In the implementation of the algorithm for reassigning TRO parameters, j-vars
were mixed with j-assigns, which is a type error.

There was no test for this case, which has been added. The algorithm is fine as
far as I can tell and test, it just needed to return JStmts, not JExprs, and do
the wrapping internally for the j-assigns rather than leaving that work until
later.