Adding support for dependency-based modified times
Build a dictionary max-dep-times that maps from each module's URI to the
MAXIMUM modified time of it and its dependents. Use this instead of just a
locator's modified time to make decisions about cache validity.

This makes it so if A imports B imports C, where B provides something that is
*calculated* from the provides of C, then A sees the update. This would not
happen previously.