Idea -- always provide all datatypes
I think this test makes clear that this is a resonable (if not the only)
option. Modules should always include the information about all datatypes
defined on them.

We _could_ do a transitive search of the provided types to find any instances
of datatypes. I don't think that's particularly interesting to do, though,
because the datatypes themselves aren't added to any namespace by this process.

The model can be that datatype information is always part of a module's static
info, but provide syntax (either the type or the data definition itself) needs
to be written down in order to make the names visible.

With this change, it's still not possible to use the type D.Data in contexts
that import provide-arrow-using-datatype, and this isn't a type error; it's
caught during resolve-scope.