Two changes: tsify build and file module updates
Add tsify to make build process use browserify and tsify together to reduce
number of build commands

Make a copy of the file.arr library and move into compiler directory so we
control it and it uses the async filelib.js. Some small fixes in filelib.js
(add flag to open(), make readFile and mtimes use paths)

readFile _has_ to use a path because browserFS appears to not allow fds in that
position. mtimes doesn't strictly have to, but since node's fs module seems to
allow paths and browserfs *prefers* them, we should move in that direction.