Committing the in-progress work on my laptop, to make sure we're in sync across both machines.
Main change here is an in-progress effort we'd made in repl.arr, to improve the names available in a provides-for-repl.  This handles include names, but doesn't yet handle including types, data, or modules.

Also, commenting out spies and adding a test case, which currently passes but which may need to be slightly tweaked once we improve the text of the shadow-id error message.