Tests: make type checker not run on certain tests
Tables can't currently be type checked, but we'd like to run tests
on them. The current testing framework forces all tests to be
type checked, meaning that any table tests we write would fail. This
commit makes the type checker ignore test files that have the string
/no-type-check/ in their filename. For example, the file
"no-type-check-table-basic-syntax.arr" will not have type checks run
on it, but the file "paren.arr" will have type checks run on it.