Add a new datatype to represent exported datatypes
Datatypes, like values, can be exported directly or can be aliased. For
example, module A may provide a datatype that is imported and re-provided by
module B, potentially at a different name. In module C, if that datatype is
imported and re-exported, the compiler needs to be able to resolve the
reference all the way back to A, which means B needs to have stored that the
datatype it provided was an alias for the one provided from A. It can't export
it as a type alias because that would lose information – in the compile context
for C, there wouldn't be enough information to tell the difference between an
alias to an underlying datatype with variants and rich structure, or a type
like { x:: Number }, making e.g. cases compilation/type checking impossible.

Now, just like values, datatype exports can express this alias relationship
across many modules. Also like values, the aliases do *not* form a linked list
in this construction. Each alias refers directly back to the originating
module, rather than requiring resolution that traverses all the modules along
the way. This reduces implementation complexity a bit; it remains to be seen if
we want the whole linked list for error reporting or not.