Add data visibility to the (non-user-visible) syntax of provides
This is necessary & useful because in resolve-scope we need a place to hang the
information to pass on. We could add this information to yet another dictionary
on a NameResolution, but rather than muck with that structure, this paves the
way for future concrete syntax, and an easily-extensible datatype that could
have meanings beyond all and opaque, all embedded in the program we generate.

This isn't being used in a meaningful way anywhere yet, just being propagated.