(For immediate revert) Make the type-checker respect opaque data
Going to revert this, but the commit is useful to comment on for posterity.

@blerner and I decided that this isn't the right feature set to *hide* data,
and we should have an explicit facility for creating opaque types, and not try
to make it solely part of the module system. This would be more in the spirit
of how ML-family languages do things, and is probably the right decision.

Trying to have implicitly-hidden data in provides ends up implicitly making a
new variety of type that has no concrete syntax, and difficult choices in how
to talk about it in error messages, render it, and discuss its definition. So
in this module system design, any context that can _refer_ to a datatype can
see it fully, including fields and using cases on it. Information can still be
hidden using closures and not exporting the datatype itself and wrapping
exported values in an object with a limited interface (if, say, it's critical
for an assignment to hide information).