Don't assume that names mentioned in provide blocks are defined
The previous behavior was that `provide: x end` would crash the compiler with
"Key x not found". This makes it be reported gracefully as an unbound id. It
also makes it so compilations that fail at this step don't produce a static
provides block, which was silly to ever do because it was nonsense if there
were scope errors