Basic support for parse/WF errors adjacent to chunks
There's a lot of UI/UX work to do here, but I think this is getting somehwere
interesting.

This exposes an endpoint called `lint` from cli-module-loader that returns
either the AST of a module or its parse/WF errors. This endpoint is called from
each of the chunks on edit (with a calming cooldown as with run), and then
any errors are propagated back to the chunk for display.

Things that are annoying right now:

- Right now compile and lint can race on an edit. We should do something like
  check that all blocks lint before calling compile. It will be work to
  coordinate that synchronization.
- The adjacent errors don't do any work to highlight within the chunk or have
  any styling.