Set CI to false to continue deploying when we see certain webpack warnings
Right now we see: Critical dependency: the request of a dependency is an expression

I've tracked it somewhat to the need to do a big update in Stopify from
babel-core to @babel/core. That's going to be an undertaking, so allowing the
build to finish with these warnings and still deploy.