Performance try – only save edits when we actually update
I did some profiling and the computation happening in React underneath setState
when currentFileContents were updated was actually pretty large. It could take
~60ms on my MacBook to do a state update in the plain text editor... and nearly
twice that on my ChromeBook.

I think in general it's OK if changing the code contents causes some relatively
expensive recomputation (we can still try to cut down). I just think we
shouldn't fire that computation on every keypress. So this commit moves the
updating of file contents to the same time as calling update() to potentially
save the file and also trigger the run/compile event.