Add out-of-module ids to create-prelude (code generation for `include`)
The reason this works is mildly subtle but (I believe) sound.

It's useful for various reasons to avoid generating a variable binding for
every imported identifier, and to instead generate them for only the imported
ids that are actually used. Since `include` is so common, there are often many
more names in scope than are used, and adding them all pollutes generated code.

The old compiler used a free-ids pass to calculate the variables that were free
from the point of view of the program body. It then used the computed
environment to figure out which uri each of these came from, and looked up the
value in the appropriate module via a dynamic dictionary stored on the runtime.

Both of these pieces—the freevars check and the module dict—are different in
this implementation.

1.  The AST pass to compute freevars is elided, in favor of using the new
module system's metadata about the origin of a definition. The compiler now
creates and fills a dictionary of all the out-of-module ids that are _used_ by
adding a dictionary update in the s-id case indicating the use. The
“new-definition” field matches this meaning exactly. Then the create-prelude
function adds a variable binding for each entry in this dictionary to the
prelude of the module via runtime lookup.

I actually started writing the freevars pass (the existing one is on the ANF
tree which doesn't exist anymore, and then realized that the new module
system's resolve-scope actually does this in all the meaningful ways we need
already. And in this case, “freevars” is actually a weird way to think about it
since all these names are bound, their binding edge just points outside the
module, which is precisely what resolve-scope computes. We may someday want a
standalone freevars pass again, but maybe not – resolve-scope is a pretty cool
module.

2. Since we're relying more on the semantics of node `require` now, we no
longer have a runtime that controls the loading of modules, so it's less clear
where to build the dynamic module dictionary. It would be a mistake to add this
to runner.ts, because makeRequire is only supposed to mimic node `require`
behavior, where this should work as well.

My first thought was to add generated code to the end of each module that would
add its provided values to the runtime before exiting. However, this is really
hard with hand-written JS modules, because the author of the module would need
to know its URI, which always sucks, because they don't necessarily know where
on the file system it is or how it might be included in the future. Instead, we
add new code to, after each _import_, add the module to the runtime module
cache. This adds a line of code for each _import_, which makes the generated
code mildly uglier in the prelude, but will allow authors of FFI modules to
avoid worrying about how their module is cached.