Make contract weaving treat tuple destruction as a blank binding
This is an attempt to fix an issue where programs like the one in the
regression test would report a duplicate contract because the arguments didn't
all show up as blank, since they were desugared to tuple annotations
(containing all blank).

I'm a little worried that this isn't the right place to do this, but it wasn't
totally clear to me if it's better to synthesize a different annotation for the
tuple destruction or not.