Fully resolve type aliases when looking them up
This enforces what we believe was an invariant before, which is that compiled
modules and handwritten JS modules need to have a local type alias for each
datatype they define. This probably could be worked around by changing the way
the type checker constructs and works with environments, but it also makes
sense to require injecting datatype names into the type environment via an
alias to keep the namespace uniform.

The main reason we added this was to canonicalize the representation of
alias-of-alias-of-alias...-of-datatype cases, where the datatype should NOT be
copied across many modules, but modules that refer to the datatype should be
able to bottom out their lookup in an alias that refers to the datatype's name
on that module.