Lift ann-to-typ and use it eagerly in resolve-scope
This works towards addressing a longstanding problem, which is that annotations
are not equi-expressive with types. In particular, there's no t-forall analog
in annotations as a syntactic choice.

This commit uses ann-to-typ eagerly in resolve-scope, instead of lazily when
generating serialized provides, in order to get forall types into the type
environment during scope resolution. This fixes the issue in the most recent
build (https://travis-ci.org/brownplt/pyret-lang/builds/627358433) where this
type was not expressible in untyped exports:

type Reducer<Acc, ....> = { .... }

This commit also makes it so we only use the `tb-none` and `tb-typ` variants of
`TypeBindTyp`, so we can simply that type.