Fix provide: x as y end when x is not also provided
This cements the meaning of a provided name's identity as the name it was
_first_ provided as across the module system. So if a name is provided at a
different name the first time, and the original definition is not provided by
name, that totally hides the original name from the rest of the module system.

This means that in this program:

provide: x, x as y end

x = 10

an importer of this module cannot import both y and x at the same name, since y
has an identity as a name originally provided from this module.

If only x were provided, and then reprovided with different names in another
module, those could be re-imported with the same name elsewhere since they have
the same underlying module uri + name identity (we do this with the image
libraries).

This could change in the future but requires yet more infrastructure for
underlying identities of definitions.