page->webworker: communicate type of message in message itself
Removes the _parley field from messages sent to the webworker,
we now rely on a "request" field whose value is the type of request.
Only messages with this field are sent to the onCompile handler
created in webworker.arr

For example, to compile a program the page sends a json object to
the webworker containing { "request": "compile-program" .... }. This
"compile-program" corresponds with the type it will be parsed as in
message.arr: Request % (is-compile-program).