Fix redux loosing track of codemirror elements
Previously, when a new chunk is created as a result of completing
an expression and hitting <enter>, the newly created chunk's
codemirror would not be properly tracked in the redux store.

The reason for this is that it used to be an uncontrolled component,
which told redux that it existed in its `editorDidMount' callback. But
since Redux avoids creating new elements whenever possible while updating
state, the new chunk would have the "old" codemirror editor.

Making our CodeMirror Controlled instead of Uncontrolled fixes this problem. It
has the added bonus that more of the chunk editor's functionality can be
controlled via Redux.