ide: Partially fix scroll jumps on chunk focus
When a CodeMirror component updates it forces the window to scroll to it. This
commit removes updates that we didn't actually need to perform, so less
scrolling happens.

The updates were for deselecting multi-chunk selections. If you select some
text, then click somewhere, the text should be deselected. We were causing
deselection updates on chunks that didn't have any selections in the first
place.

This commit does not remove *all* deselection updates; chunks that are actually
selected will still be deselected. This means that focus will still jump if the
selected chunks are scrolled out of screen at the time of clicking.

It would be useful to investigate why the window scrolls when an off screen
CodeMirror updates---maybe we should try to disable this.