make type check errors use the to-string() method directly
It looks like constructor printing used to be overridden for the sake
of errors. However, this is silly because we want constructor printing
for spying on types and so on within the compiler.

So make the error rendering contexts use to-string explicitly.

Down the line this should be a richer renderable type struct, but
ED.embed does constructor printing as well so stringifying with
the existing stringifier is better.