Remove chunk focus on menu dropdown click
The dropdown is hidden when the "on blur" event is fired. This happens whenever
another element is focused. If a chunk is focused before the dropdown is
clicked, that focus will be asserted on the redux update that makes the dropdown
visible. This instantly triggers the on blur event, causing the dropdown to hide
itself.