Add some checking to raw-array-build
1. Make sure fractional sizes are not allowed
2. Make sure negative sizes are not allowed
3. Make sure sizes greater than what we can allocate in JS are not allowed

For 1, we can know this isn't affecting any programs that expect behavior like
floor(), because the current behavior treated fractions like 0 so nothing would
be allocated.

Ditto for 2.

For 3, I can see arguments for another (larger) size, potentially, but also I
don't think it's unreasonable for different platforms to have different limits,
and this is a pretty reasonable limit for JS. As in, I don't think we're going
to implement massive arrays atop our JS implementation, and if the language
moves on to another platform it can have a different limit there.