Several small fixes to tests and environment exposed by fixing #1505
It turns out that the last check block in a program could end up not visited by
the type-checker.

Altering
https://github.com/brownplt/pyret-lang/commit/3e6b258703ab074276b8e1f80df32cdf15fba251
makes these always get visited (in addition to cleaning up scope).

This makes a few necessary cleanups apparent in the type-check tests:

- There was a test that used bare StringDict, which wouldn't unify with an
  inferred instantiated StringDict<?> type on the RHS of a binding
- There were test operators missing from global.js, making some of the
  lesser-used testing operators fail to type-check correctly

This also fixes a bug in pretty-printing s-module-ref that I encountered while
printing the intermediate ASTs to debug this.