Sketch of extend using getter infrastructure
Prototype inheritance is quite tricky here! Use a full copy for now.

The issue with newObj = Object.create(obj) is that the method isn't in the
method cache yet, so there ends up being no method property as an *own*
property of newObj, which breaks looking for the getter function at the
beginning of the getter to cache.