Two housekeeping changes in building up the page
1. Plumb the `pipeline` option through to the messages passed from page to
webworker (left as `anchor` for now)
2. When `recompile-builtins` is false (the setting used on the page to avoid
unnecessary file access) also don't *save* the builtin files if they already
exist. Since this tells the compiler to trust the explicit in-memory cache,
this is strictly silly to do. The lack of this check was making it so files
could get copied on startup, updating mtimes and causing extra Stopification of
builtin modules (like list!), slowing down first run significantly.