Make direct-codegen *also* use is-, and revert a change to type-check
There was a compilation type error in sets with the type change to the is-Data
types, I'm not quite sure why. It was reporting something about right/is-leaf(right) in

  fun greatest(t):
    cases(AVLTree) t:
      | leaf => raise("Went too far in traversal step")
      | branch(_, _, _, right) => if is-leaf(right): t else: greatest(right) end
    end
  end

But I thought what I'd changed would only affect is-AVLTree. Reverting this TC
change for now, leaving in the renaming to is- and making it consistent