Fix several path-related issues that manifested in the web interface
1. .realpath() in BrowserFS seems to not do the right thing (?) .resolve()
seems to be the right thing. Use that.

2. In getBaseDir when buildPath and source had an identical prefix and never
differed, we were not cutting anything off, which was wrong. So set
`cutoffIndex` to the length if we don't find a differing spot in the loop.

3. In node, the '.js' extension is at least optional, and maybe
required/generated by some TS stuff. So make our requires allow for/do that as
well.