Deserialize chunks into marks
There may be some performance implications from saving all the time. Not sure
if that's the right protocol. But for the most part, this works decently
well. You'll note i used a pattern that's essentially what a Controlled
component would do. However, react-codemirror2 is kind of a mess, and was
giving me bad bad values, so i rolled my own. It's nicer working with an editor
instance than a string *anyway*