Allow shadowing included ids, and more configurable REPL
1. A new check in resolve-scope before reporting shadowing errors that doesn't
count shadowing errors if they are between an id and an included id (except in
the case of REPL ids from definitions:// and interactions://, which are special
origins). This is bad in some ways because e.g. the beginner who writes `circle
= circle(50, "solid", "red")` will likely be confused. But it's worth seeing
what this looks like.

2. Allow clients of the repl library to pass in preferred extra imports when
constructing the definitions locator.

3. Allow ExtraImport s to be expressed as blanket include statements