Progress, not complete, on debugging stopify + reactors
- Do a hacky version of @blerner's suggestion to instantiate stopify once and
  thread through (this can and needs to be refactored a bit I think)
- Expose `getCurrentRunner()` from runner.ts so other contexts can use it

This commit BREAKS reactors for the moment, but unbreaks several other things,
like re-running programs that use interesting runtime functions, so it's a
useful checkpoint. The next step is to use the evalCompiled strategy in
runner.ts's run method and the adaptWorldConfig function to have a
consistent way of doing these runs everywhere.