Stopify wrappers in reasonable places
All of the /* @stopify flat */ here may not be necessary, but I want to get a
good checkpoint in (and review with someone)

This commit does two things:

1. Use stopifyRunner.runStopifiedCode to run thunks that may (transitively)
call code that may have been stopified. It's there, it works, and it avoid some
shenanigans we've been doing.

2. Add a new runtime function safeVoidCallback. safe because it wraps the run
in a stopify context, VoidCallback because it assumes it's getting a callback
so has no reason to return to anyone. This is used to wrap callbacks that will
be called in a completely blank context by the browser in e.g. event handling.

Here's what was happening: world-lib.js was stopified, so *any of its code, at
any time* could throw a suspend or other stopify control exception! This led to
unpredictable timing/etc errors where *the wrapper code around handling ontick*
would have issues.

This moves the stopify wrappers to the point of registering the callback (hence
a callback wrapping helper), so the context gets immediately created *around*
the function (not even inside it).

Another solution to this is probably to make world-lib.js not get stopified.
But if it doesn't its module isn't in quite the right format, so it gets tricky.