Change text mode to use new async style runner
The three handlers for text mode, run(success/failure)/compilefailure on
Programs (as opposed to segments) are gratifyingly small and use existing
infrastructure.

This is in preparation for hopefully deleting a bunch of gnarly code and moving
more completely over to the new session-based infrastructure for running.

Also some renames – for stuff that runs all the chunks/segments, call it
runSegments, and for the stuff that runs the whole program, call it runProgram