include this in pyret-lang, not just in cpo
This is actually a bit of an issue. CPO is going to expose stuff that isn't
available at the command line. We cannot write essentials2021 in the same way
we can on CPO because image isn't available in the same way.

So we either need parallel essentials20XX packages in the lang vs. CPO, or
define everything in pyret-lang and stub out the stuff that won't work on CPO.
Needs at least a little thought, and maybe thought on naming, to make it work.