More undo/redo, but second thoughts about model
Another problem that comes up is you have to be careful that the events you
put to future are *updated*, so that if you undo deleting a chunk, then modify
it, then redo deletion, then undo it again, you get the modified text rather
than the old text. This is a very dangerous bug!!!

In many undo systems, you don't have to worry about this because all important
state is tracked through the undo system, but in ours we have many separate
undo zippers, so it's more delicate.

We could also consider trying to share undo between our codemirrors and our
chunks, i think it might be possible, and threading undo through them all. i'm
not sure whether that's what we *want* though