Mechanics of the stop button
A few things internal to stopify seem odd or badly implemented by Joe's
evalCompiled. This commit does a few things to massage Stopify's runtime into a
state where it can end incremental runs and start the next one cleanly.

Our goal is to have the stop button pause and end the current run with an
exception. We tried to do that by calling onEnd on the runner, but this doesn't
set enough state to clean things up, nor is the onDone handler set correctly by
evalCompiled.

One option we didn't pursue because we've had issues in the past is
reinstantiating the stopify runner every time. We don't want to do this because
we cache instantiated modules from Pyret's runtime and stdlib, and don't want
to think about those modules potentially closing over other stopify
runtimes/runners.

Instead, we carefully are setting state to get to a clean restart (we think) in
between runs, even if they end in a manually-thrown exception using onEnd.

The bits of state we need to set up are:

- Setting the onDone handler in evalCompiled so we get per-eval ending
  callbacks (the call to runInit)
- Setting the yield flags back to their initial state of { kind: 'resume' }
  manually, which is horrifying and unsafe

The former change should probably be a part of evalCompiled, because of how
onDone is handled (though it's not totally clear that there is only ever one
onDone to consider, for example if the program evalCompiled calls does various
kinds of pauseK and another evalCompiled nested...).

The latter change should probably be either a part of runInit or a part of
onEnd, to either set up or cleanup, whichever makes the most sense.