Don't duplicate errors, link between highlights
You can click a highlight to scroll its referred chat into view. The referred
chat no longer has its results blown away (those results are not only still
relevant, they're still correct!), instead an "incoming envelope emoji" is
displayed to it left by the insert button. That envelope can also be pressed,
scrolling the erroring chat into view.

TODO: Clicking the reference letter to scroll down scrolls the relevant
CodeMirror into view, all the way to the bottom, but does not focus the error
message like it definitely should. This is because it's easy to scroll the
CodeMirror into view because we already have CM instances laying around in
Redux. The right way to do this is to put some refs into state somewhere,
hopefully of the highlight content itself; a hack that might work is to mess
around in the DOM after getting editor.getWrapperElement() to get its lower
sibling and assume that's the error

Accessibility note: each of these only scrolls the relevant section into
view. It's much harder to figure out how to give actual browser focus to that
section, but it should definitely be done. Mostly, it should be done after
fixing the above so the correct element is focused.