Change runners to have a separate initial state update helper
This makes it so the check for `running` not being idle hits correctly. Before,
the asynchronous state update of running in the individual runners could
trigger only after a second run came in and started. This could cause conflicts
between two different versions of the chunk.

Notably, this was triggered on deleting a chunk which could trigger two runs.

It's going to be hard (esp. including user interactions) to avoid two events
*trying* to start a run, so we guard it and this ended up being a good test
case of that guard.