make `include` and `context` uses of sets.arr happier with the type checker
The main thing here is that the types are still in type-defaults.arr. But the
type-checker now wants more information about datatypes, and modules are more
stringent about what it means to provide `data` (e.g. the variants and friends
are included). These were out of sync.

Fix this by providing the right things from the module AND listing them in
type-defaults.arr.