Remove s-ref from AST
s-ref was part of an early design of mutable fields that could represent
a ref as a standalone expression.

This hasn't been possible in surface syntax for some time, and given
that the design of ref fields is working in PAPL/DCIC, it seems
unlikely that we need to maintain vestigial syntactic support for the
idea.