Add wrapping fix from runner.ts for precompiled code
This fix had already been applied with wrapContent in runner.ts, but not
applied to pre-stopified code. The comment in runner.ts describes this

https://github.com/brownplt/pyret-lang/blob/d29d95d9bda9a36b877dca9e61ffd963cfc51826/src/webworker/runner.ts#L70

The aliasing trick wasn't actually working in a specific case that mattered for
the IDE. If `global.arr.js` was already dynamically cached, it would never load
and trigger the global variable require to use /compiled/builtin as the right
cwd. Instead it would use to the cwd of the original project, and fail to find
option. This is the more robust fix to globals getting changed due to stopify
and forcing them to be variable references