Fix #1643 by not having include include data
We had a nice design discussion about this in the past that data definitions
need to be present on all modules so that the type-checker can see them.

Then, provide/include `data` just means sugar for a bunch of type/value imports.

This actually can be made to work really well for just providing the data type
without its variants – the main solution is just to have `include foo` _not_
try to do `include from foo: ... data * end` as well as `type *` and `*`. In
fact, `data *` is always redundant for a blanket include (or it would introduce
errors looking for things that aren't present).

Arguably `provide *` would potentially even be better defined as `provide: type
*, * end` because the type part will pick up the aliases to the datatypes
anyway, and the `*` part will pick up all the functions.

This commit at least (a) passes all the test we have (b) passes a related
regression and (c) doesn't break the teachpacks that use chart-internal variant
names.