Changes for updating/cleaning valueskeleton
There are a few main parts:

- Update valueskeleton to more closely mirror all the Pyret values, but with
  VS replacing any pyret value.
- Change the signature of _output so that it takes a recursive to-output
  function (just like equality).
- Adapt the existing to-repr algorithm to just generate valueskeletons fully
  eagerly, calling into _output directly when datatype-defined renderes are
  present

This is based on a few observations:

1. We haven't really observed cases where just calculating the valueskeleton
for a datum is problematic, just rendering. A lot of the complexity of the
previous implementation was in making some lazy patterns in visiting the
value possible, but everything we did was effectively eager anyway.

2. We can simplify all the reentrant/stack management stuff dramatically

3. By representing more things directly in valueskeleton, we remove some
reliance on callbacks between the runtime and the rendering process (e.g. the
page), which will never be totally simple due to e.g. stopify. Better to have
fewer coroutines.

The recursive to-output is important because it closes over the cache that
checks for cyclic occurrences; folks implementing their own _output methods
need to be aware of this (_output isn't really documented, but now we know what
to say)

For backwards compatibility, we may want to rename this method to e.g.
_to-output, leaving the old one intact to not break code out in the wild. We
can have vs-value be an alias for to-output, and smart constructors for
anything that used to work with lists but no longer does. This will only break
(maybe) some esoteric static type-checking cases old uses of _output, and users
writing a a custom to-output and relying on Pyret to detect it as part of a
cycle. I'd be surprised if that exists in the wild, and the benefit is that we
get something much, much simpler to work with.