Fix a type-checker issue related to mutably updating variant fields
The key change is mapping + deep copying tuples representing field names and
values for variants when substituting.

This manifested as a mutable update to aliased fields like the "val" field of
an Option or a Pick. It made a free existential be the apparent type of those
fields because the type checker would change the original type rather than
copying the type "schema" and substituting into it for each instantiation.

The tests are all small tests that would fail in the presence of the bug.

The change to ast-util is *intended* to catch similar situations in the future,
but isn't catching this specific error, which we don't quite have a dynamic
assertion for yet.