Refactoring RangeWidget + friends to be less typey-wobbly
This still isn't great. It basically took the `instanceof ContainerRange` and
stuck it into ArrayWidget and RangeBoxesWidget rather than having to include it
in ValueSkeleton somehow.

RangeWidget is like a fancy summarize-able list widget that needs to show an
expander when its immediate parent is expanded and not otherwise, but is always
summarized to an array range. It ends up mixed up, via splitIndexRange and
abbreviate, in a list adjacent to ValueSkeletons. This should probably be
refactored into some kind of [VS[], RangeWidget, VS[]] triple for RangeBoxes
and some kind of RangeWidget[] for the expanded case. It's tough to keep track
of the difference.

This commit has another refactoring for ArrayWidget that isn't used yet, but is
there in preparation for vs-record, which is incorporating key/value pairs. It
might be worth un-doing this since the complexity it adds isn't great, and it's
not clear we even want the same expansion/contraction rules for vs-constr as we
do for collections.

But, it's useful for reconstructing vs-constr on top of the old system, so
leaving it in for review + discussion.