First round of refactoring builtins to a single directory with ordering in Makefile rule
Some fun discoveries of cycles:

- string depends on list (because of e.g. string-explode)
- error depends on string (because of course)
- list (wants to) depend on error

We can cheat with string and use dynamic requires to get string to not require
that lists be built first, which may just be a bootstrapping reality.

Also, we have an issue that global depends on list, but we've been using global
as a de facto "initial environment" in a lot of places, including in generated
builtin URLs in the compiler. So I think some sneaky cycles/missed URI lookups
will happen here.

I tracked down a bunch by redirecting things to primitive-types and using
runtime-globals more, but more work remains.

Still, this is good simplifying progress and exposing things that weren't
obvious with the previous setup.