Be more careful about loading all scripts in order
I'm pretty sure the issue we saw today was a race between client.js and
api-wrapper -- api-wrapper starts with a call to reauth, and both errors were
on the call to setToken on line 68 in api-wrapper. That would be consistent
with that JavaScript running before the google client library loaded.

So, make these (annoyingly) wait for one another. This needs to be rewritten
with plain Promise and its own calls to the /refreshAccess endpoint.