UI updates for prompt positioning
1. Move prompt and its hint inside chat scroll
2. Make chats list not flex-grow, so that the prompt sticks to the bottom
3. Add some margin below the prompt to allow nudging it up on screen

I'm not sure what the flex-direction: column-reverse was doing here, but given
that the grid containing prompt and chat-scroll isn't really being used
anymore, I think this is now doing exactly the right thing.

This moves the prompt up to right below the last entry. It no longer floats
above all the program textwhen you scroll up; you have to scroll back down to
get to it.

A further refinement would be to try to make it so it sticks at the bottom of
the screen when scrolling up, but then gets some margin when you scroll all the
way down. I'm not sure which is better; I feel like once the program gets long
enough, you don't want to blindly type in new prompt-stuff anyway so scrolling
down for context is useful.